backend/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
ebooklib
beautifulsoup4
python-dotenv
supabase